Move workspace terminal page into its own file
