+++ app/src/main/java/me/rerere/rikkahub/ui/pages/extensions/WorkspaceTerminalPage.kt
@@ -1,0 +1,291 @@
+package me.rerere.rikkahub.ui.pages.extensions
+
+import android.graphics.Typeface
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.core.content.res.ResourcesCompat
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.termux.terminal.TerminalSession
+import com.termux.view.TerminalView
+import me.rerere.rikkahub.R
+import me.rerere.rikkahub.ui.components.nav.BackButton
+import org.koin.androidx.compose.koinViewModel
+import org.koin.core.parameter.parametersOf
+
+@Composable
+fun WorkspaceTerminalPage(id: String) {
+    val vm: WorkspaceDetailVM = koinViewModel(parameters = { parametersOf(id) })
+    val state by vm.state.collectAsStateWithLifecycle()
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(
+                        text = state.workspace?.name?.let { "$it · 终端" } ?: "终端",
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis,
+                    )
+                },
+                navigationIcon = { BackButton() },
+                colors = TopAppBarDefaults.topAppBarColors(
+                    containerColor = Color.Black,
+                    scrolledContainerColor = Color.Black,
+                    navigationIconContentColor = Color.White,
+                    titleContentColor = Color.White,
+                    actionIconContentColor = Color.White,
+                ),
+            )
+        },
+        containerColor = Color.Black,
+    ) { innerPadding ->
+        WorkspaceTerminalContent(
+            root = state.workspace?.root,
+            contentPadding = innerPadding,
+        )
+    }
+}
+
+@Composable
+private fun WorkspaceTerminalContent(
+    root: String?,
+    contentPadding: PaddingValues,
+) {
+    val context = LocalContext.current
+    val terminalTextSizePx = with(LocalDensity.current) { 13.sp.roundToPx() }
+    val terminalTypeface = remember(context) {
+        ResourcesCompat.getFont(context, R.font.jetbrains_mono) ?: Typeface.MONOSPACE
+    }
+    var finished by remember(root) { mutableStateOf(false) }
+    var controlDown by remember(root) { mutableStateOf(false) }
+    var altDown by remember(root) { mutableStateOf(false) }
+    val sessionClient = remember(root) {
+        WorkspaceTerminalSessionClient(context.applicationContext) {
+            finished = true
+        }
+    }
+    val viewClient = remember(root) {
+        WorkspaceTerminalViewClient(context)
+    }
+    viewClient.controlDown = controlDown
+    viewClient.altDown = altDown
+
+    if (root == null) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(contentPadding)
+                .padding(16.dp),
+            contentAlignment = Alignment.Center,
+        ) {
+            Text(
+                text = "工作区加载中...",
+                color = Color.White.copy(alpha = 0.7f),
+            )
+        }
+        return
+    }
+
+    if (!workspaceRootfsReady(context, root)) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(contentPadding)
+                .padding(16.dp),
+            contentAlignment = Alignment.Center,
+        ) {
+            Text(
+                text = "请先安装 Rootfs",
+                style = MaterialTheme.typography.bodyLarge,
+                color = Color.White.copy(alpha = 0.7f),
+            )
+        }
+        return
+    }
+
+    val session = remember(root) {
+        createWorkspaceTerminalSession(context, root, sessionClient)
+    }
+
+    DisposableEffect(session) {
+        onDispose {
+            sessionClient.terminalView = null
+            viewClient.terminalView = null
+            session.finishIfRunning()
+        }
+    }
+
+    Surface(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(contentPadding)
+            .imePadding(),
+        color = Color.Black,
+    ) {
+        Column(modifier = Modifier.fillMaxSize()) {
+            Box(
+                modifier = Modifier
+                    .weight(1f)
+                    .fillMaxWidth(),
+            ) {
+                AndroidView(
+                    modifier = Modifier.fillMaxSize(),
+                    factory = { viewContext ->
+                        TerminalView(viewContext, null).apply {
+                            isFocusable = true
+                            isFocusableInTouchMode = true
+                            setTextSize(terminalTextSizePx)
+                            setTypeface(terminalTypeface)
+                            setTerminalViewClient(viewClient)
+                            attachSession(session)
+                            sessionClient.terminalView = this
+                            viewClient.terminalView = this
+                            setOnTouchListener { _, event ->
+                                if (event.action == android.view.MotionEvent.ACTION_UP) {
+                                    viewClient.focusAndShowKeyboard()
+                                }
+                                false
+                            }
+                            post {
+                                viewClient.focusAndShowKeyboard()
+                            }
+                        }
+                    },
+                    update = { terminalView ->
+                        terminalView.isFocusable = true
+                        terminalView.isFocusableInTouchMode = true
+                        terminalView.setTextSize(terminalTextSizePx)
+                        terminalView.setTypeface(terminalTypeface)
+                        terminalView.setTerminalViewClient(viewClient)
+                        sessionClient.terminalView = terminalView
+                        viewClient.terminalView = terminalView
+                        terminalView.setOnTouchListener { _, event ->
+                            if (event.action == android.view.MotionEvent.ACTION_UP) {
+                                viewClient.focusAndShowKeyboard()
+                            }
+                            false
+                        }
+                        terminalView.attachSession(session)
+                        terminalView.onScreenUpdated()
+                    },
+                )
+                if (finished) {
+                    Text(
+                        text = "终端已退出",
+                        modifier = Modifier
+                            .align(Alignment.BottomCenter)
+                            .padding(12.dp),
+                        style = MaterialTheme.typography.bodySmall,
+                        color = Color.White.copy(alpha = 0.7f),
+                    )
+                }
+            }
+            TerminalExtraKeysBar(
+                controlDown = controlDown,
+                altDown = altDown,
+                onControlToggle = { controlDown = !controlDown },
+                onAltToggle = { altDown = !altDown },
+                onSendText = { session.writeText(it) },
+            )
+        }
+    }
+}
+
+@Composable
+private fun TerminalExtraKeysBar(
+    controlDown: Boolean,
+    altDown: Boolean,
+    onControlToggle: () -> Unit,
+    onAltToggle: () -> Unit,
+    onSendText: (String) -> Unit,
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(Color.Black)
+            .horizontalScroll(rememberScrollState())
+            .padding(horizontal = 8.dp, vertical = 6.dp),
+        horizontalArrangement = Arrangement.spacedBy(6.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        TerminalExtraKey("ESC") { onSendText("\u001B") }
+        TerminalExtraKey("TAB") { onSendText("\t") }
+        TerminalExtraKey("CTRL", selected = controlDown, onClick = onControlToggle)
+        TerminalExtraKey("ALT", selected = altDown, onClick = onAltToggle)
+        TerminalExtraKey("-") { onSendText("-") }
+        TerminalExtraKey("/") { onSendText("/") }
+        TerminalExtraKey("|") { onSendText("|") }
+        TerminalExtraKey("←") { onSendText("\u001B[D") }
+        TerminalExtraKey("↓") { onSendText("\u001B[B") }
+        TerminalExtraKey("↑") { onSendText("\u001B[A") }
+        TerminalExtraKey("→") { onSendText("\u001B[C") }
+        TerminalExtraKey("HOME") { onSendText("\u001B[H") }
+        TerminalExtraKey("END") { onSendText("\u001B[F") }
+    }
+}
+
+@Composable
+private fun TerminalExtraKey(
+    label: String,
+    selected: Boolean = false,
+    onClick: () -> Unit,
+) {
+    Text(
+        text = label,
+        modifier = Modifier
+            .background(
+                color = if (selected) {
+                    MaterialTheme.colorScheme.primary
+                } else {
+                    Color.White.copy(alpha = 0.12f)
+                },
+                shape = RoundedCornerShape(6.dp),
+            )
+            .clickable(onClick = onClick)
+            .padding(horizontal = 12.dp, vertical = 8.dp),
+        style = MaterialTheme.typography.labelMedium,
+        color = if (selected) {
+            MaterialTheme.colorScheme.onPrimary
+        } else {
+            Color.White.copy(alpha = 0.9f)
+        },
+    )
+}
+
+private fun TerminalSession.writeText(text: String) {
+    val bytes = text.toByteArray()
+    write(bytes, 0, bytes.size)
+}
